+++ DelvUI/Interface/Party/PartyFramesHud.cs
@@ -3,15 +3,12 @@ using DelvUI.Config;
 using DelvUI.Enums;
 using DelvUI.Helpers;
 using DelvUI.Interface.GeneralElements;
-using FFXIVClientStructs.FFXIV.Client.Game.Character;
 using ImGuiNET;
-using Lumina.Excel.GeneratedSheets;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
 using System.Runtime.InteropServices;
-using static System.Net.Mime.MediaTypeNames;
 
 namespace DelvUI.Interface.Party
 {
@@ -34,7 +31,7 @@ namespace DelvUI.Interface.Party
         private bool _layoutDirty = true;
 
         private readonly List<PartyFramesBar> bars;
-        private LabelHud _labelHud;
+        private LabelHud _titleLabelHud;
 
         private bool Locked => !ConfigurationManager.Instance.IsConfigWindowOpened;
 
@@ -57,7 +54,7 @@ namespace DelvUI.Interface.Party
                 bars.Add(bar);
             }
 
-            _labelHud = new LabelHud(config.ShowPartyTitleConfig);
+            _titleLabelHud = new LabelHud(config.ShowPartyTitleConfig);
 
             PartyManager.Instance.MembersChangedEvent += OnMembersChanged;
             UpdateBars(Vector2.Zero);
@@ -93,7 +90,6 @@ namespace DelvUI.Interface.Party
             Configs.HealthBar.ValueChangeEvent -= OnLayoutPropertyChanged;
             Configs.HealthBar.ColorsConfig.ValueChangeEvent -= OnLayoutPropertyChanged;
             PartyManager.Instance.MembersChangedEvent -= OnMembersChanged;
-            _labelHud.Dispose();
         }
 
         private void OnMovePlayer(PartyFramesBar bar)
@@ -445,7 +441,8 @@ namespace DelvUI.Interface.Party
 
             AddDrawAction(Config.ShowPartyTitleConfig.StrataLevel, () =>
             {
-                _labelHud.DrawText(origin + Config.Position, TextTagsHelper.GetPartyListTitle());
+                Config.ShowPartyTitleConfig.SetText(PartyManager.Instance.PartyTitle);
+                _titleLabelHud.Draw(origin + Config.Position);
             });
         }
     }
@@ -495,6 +492,7 @@ namespace DelvUI.Interface.Party
                 ConfigurationManager.Instance.GetConfigObject<PartyFramesCooldownListConfig>()
             );
         }
+
 
     }
     #endregion